BpzCore/src/main/java/lets/develop/BpzCore/entity/Proveedor.java
package lets.develop.BpzCore.entity;

import java.io.Serializable;

public class Proveedor implements Serializable{

	private static final long serialVersionUID = 1L;
	private Integer id;
	private String nombre;
	private String estado;
	

	public Proveedor() {
		super();
	}

	public String getEstado() {
		return estado;
	}

	public void setEstado(String estado) {
		this.estado = estado;
	}

	public Integer getId() {
		return id;
	}
	public void setId(Integer id) {
		this.id = id;
	}
	public String getNombre() {
		return nombre;
	}
	public void setNombre(String nombre) {
		this.nombre = nombre;
	}
}
